fix(migrations): cast relevance literals to REAL in search_documents RPC
CASE expression in search_documents produced 'numeric' type by default but
the function declared the return column as REAL. PostgreSQL strict type
checking caused PostgREST to return {"code":"42804"} on RPC calls.

Fix: explicit ::REAL casts on all CASE branch literals (1.0::REAL, 0.5::REAL).
Also applied live via psql + NOTIFY pgrst 'reload schema' to hot-reload cache.

+++ infra/docker/migrations/002_postgrest_views.sql
@@ -195,8 +195,8 @@ BEGIN
         u.full_name as uploaded_by_name,
         -- Simple relevance scoring based on title match
         CASE 
-            WHEN LOWER(d.title) LIKE LOWER('%' || search_term || '%') THEN 1.0
-            ELSE 0.5
+            WHEN LOWER(d.title) LIKE LOWER('%' || search_term || '%') THEN 1.0::REAL
+            ELSE 0.5::REAL
         END as relevance
     FROM documents d
     LEFT JOIN users u ON d.uploaded_by = u.id
